moved help into src/modules/commands. Join commands descriptions with map instead
